Split custom input dialogs to separate files
